% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genotypes}
\alias{example_genotypes}
\title{Example of genotypic data.frame}
\format{An object of class \code{data.frame} with 100 rows and 3333 columns.}
\source{
These data are simulated and was generated using the serious game
  "PlantBreedGame". See: \url{https://github.com/timflutre/PlantBreedGame}
}
\usage{
example_genotypes
}
\description{
A data frame with genotypic data encoded in allele doses: 100 individuals (from specie \link[breedSimulatR]{example_Specie}) and
3333 markers (see: \link[breedSimulatR]{example_SNPs}).
}
\keyword{datasets}
