% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pop}
\alias{example_pop}
\title{Example of population object}
\format{An object of class \code{population} (inherits from \code{R6}) of length 11.}
\usage{
example_pop
}
\description{
Example of population object representing 100 individuals (from specie
\link[breedSimulatR]{example_Specie}) the genotypes of these individuals came
from \link[breedSimulatR]{example_genotypes} data.
}
\keyword{datasets}
