% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_SNPs}
\alias{example_SNPs}
\title{Example of SNPinfo object}
\format{An object of class \code{SNPinfo} (inherits from \code{R6}) of length 10.}
\usage{
example_SNPs
}
\description{
Example of SNPinfo object representing SNP information about the
\link[breedSimulatR]{example_genotypes} data.
}
\keyword{datasets}
