% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Specie}
\alias{example_Specie}
\title{Example of Specie object}
\format{An object of class \code{Specie} (inherits from \code{R6}) of length 12.}
\source{
These data are based on simulated data from the serious game
  "PlantBreedGame". See: \url{https://github.com/timflutre/PlantBreedGame}
}
\usage{
example_Specie
}
\description{
Example of Specie object representing a fake diploid specie "Statisticae
exempli" with 10 chromosome of length 10^6.
}
\keyword{datasets}
