% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype.R
\name{haplotype}
\alias{haplotype}
\title{R6 Class representing an haplotype}
\description{
haplotype object store specific information of individuals haplotype
}
\examples{

## ------------------------------------------------
## Method `haplotype$new`
## ------------------------------------------------

# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     lchrCm = 100,
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       physPos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       linkMapPos = NA,
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)
# simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                   nrow = 2)
colnames(rawHaplo) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SNPinfo}}{[SNPinfo class] information about the haplotype of the
SNPs (see:\link[breedSimulatR]{SNPinfo})}

\item{\code{values}}{[list] list haplotype values}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{allelDose}}{[numeric] vector of genotypes encoded in allele dose}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{haplotype$new()}}
\item \href{#method-clone}{\code{haplotype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Haplotype object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{haplotype$new(SNPinfo, haplo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SNPinfo}}{[SNPinfo class] information about the haplotype's SNPs
(see:\link[breedSimulatR]{SNPinfo})}

\item{\code{haplo}}{[matrix] named matrix of the haplotypes for all markers. This matrix should have `SNPinfo$specie$ploidy` line and `SNPinfo$nSNP()` columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `haplotype` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     lchrCm = 100,
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       physPos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       linkMapPos = NA,
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)
# simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                   nrow = 2)
colnames(rawHaplo) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{haplotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
