% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{individual}
\alias{individual}
\title{R6 class representing an individual}
\description{
individual object store specific information about one individual
}
\examples{

## ------------------------------------------------
## Method `individual$new`
## ------------------------------------------------

# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     recombRate = 0.006,
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)
# simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE), nrow = 2)
colnames(rawHaplo) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo)
myInd <-  individual$new(name = "Ind 1",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo,
                         verbose = FALSE)

## ------------------------------------------------
## Method `individual$generateGametes`
## ------------------------------------------------

myInd$generateGametes()
myInd$generateGametes(2)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{[string] Name of the individual}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[breedSimulatR]{specie})}

\item{\code{parent1}}{[string] Name of the individual's parent}

\item{\code{parent2}}{[string] Name of the individual's parent}

\item{\code{haplo}}{[haplotype class] Haplotype of the individual (see: \link[breedSimulatR]{haplotype})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{individual$new()}}
\item \href{#method-generateGametes}{\code{individual$generateGametes()}}
\item \href{#method-clone}{\code{individual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new individual object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$new(
  name = "Unnamed",
  specie = specie$new(),
  parent1 = NA,
  parent2 = NA,
  haplo = NA,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{[string] name of the individual}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[breedSimulatR]{specie})}

\item{\code{parent1}}{[string] Name of the individual's parent}

\item{\code{parent2}}{[string] Name of the individual's parent}

\item{\code{haplo}}{[haplotype class] Haplotype of the individual (see:
\link[breedSimulatR]{haplotype})}

\item{\code{verbose}}{[boolean] display information}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `individual` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     recombRate = 0.006,
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)
# simulate haplotype
rawHaplo <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE), nrow = 2)
colnames(rawHaplo) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo)
myInd <-  individual$new(name = "Ind 1",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo,
                         verbose = FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateGametes"></a>}}
\subsection{Method \code{generateGametes()}}{
Get the number of SNPs per chromosomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$generateGametes(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{[float] number of gemetes to create (default: 1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of gametes. A gamete is a named vectors with value 0 or 1.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{myInd$generateGametes()
myInd$generateGametes(2)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{individual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
