% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{makeCrosses}
\alias{makeCrosses}
\title{Proceed several crosses}
\usage{
makeCrosses(crosses, pop)
}
\arguments{
\item{crosses}{data.frame with crossing instructions: parents names
\code{ind1} \code{ind2}, number of descendant \code{n} and names of
descendant \code{names}}

\item{pop}{list of individuals containing the parents}
}
\value{
list of new individuals
}
\description{
Proceed several crosses
}
\examples{
# create specie:
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     recombRate = 0.006,
                     verbose = FALSE)

# simulate SNP:
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object:
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)


# simulate haplotype:
rawHaplo1 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                    nrow = 2)
colnames(rawHaplo1) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo1 <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo1)
rawHaplo2 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                    nrow = 2)
colnames(rawHaplo2) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo2 <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo2)
rawHaplo3 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE),
                    nrow = 2)
colnames(rawHaplo3) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo3 <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo3)


# create individuals:
myInd1 <-  individual$new(name = "Ind 1",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo1,
                         verbose = FALSE)
myInd2 <-  individual$new(name = "Ind 2",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo2,
                         verbose = FALSE)
myInd3 <-  individual$new(name = "Ind 3",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo3,
                         verbose = FALSE)
myPop <- population$new(name = "My Population 1",
                        inds = list(myInd1, myInd2, myInd3),
                        verbose = FALSE)

crossToDo <- data.frame(ind1 = c("Ind 1", "Ind 1", "Ind 2"),
                         ind2 = c("Ind 2", "Ind 3", "Ind 3"),
                         n = 1,
                         names = c("Off 1-2", "Off 1-3", "Off 2-3"))

makeCrosses(crossToDo, myPop)
}
