% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{makeSingleCross}
\alias{makeSingleCross}
\title{Cross two individuals together}
\usage{
makeSingleCross(ind1, ind2, n = 1, names = NULL, verbose = T)
}
\arguments{
\item{ind1}{parent 1}

\item{ind2}{parent 2}

\item{n}{number of descendants}

\item{names}{names of the descendants}

\item{verbose}{print informations}
}
\value{
list of new individuals
}
\description{
Cross two individuals together
}
