% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotyper}
\alias{phenotyper}
\title{R6 class representing a phenotyper}
\description{
phenotyper object is used to phenotype populations
}
\examples{

## ------------------------------------------------
## Method `phenotyper$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                    lchr = 10^6,
                    lchrCm = 100,
                    specName = "Geneticae Exempli",
                    ploidy = 2)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                   specie = mySpec)
example_pop <- createPop(geno = exampleData$genotypes,
                        SNPinfo = SNPs,
                        popName = "Example population")
myTrait1 <- trait$new(name = "Trait1",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.3))
myTrait2 <- trait$new(name = "Trait2",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.4))

phenoLab1 <- phenotyper$new(name = "My phenoLab",
                            traits = list(myTrait1, myTrait2),
                            plotCost = 1,
                            mu = c(100, 50),
                            ve = NULL,
                            he = c(0.4, 0.55),
                            pop = example_pop)

phenoLab2 <- phenotyper$new(name = "My phenoLab",
                           traits = list(myTrait1, myTrait2),
                           plotCost = 2,
                           mu = c(100, 50),
                           ve = c(9, 7),
                           he = NULL,
                           pop = NULL)


## ------------------------------------------------
## Method `phenotyper$trial`
## ------------------------------------------------


pheno1 <- phenoLab1$trial(example_pop, rep = 4, offset = c(3, 5))
pheno1$cost
summary(pheno1$data)

pheno2 <- phenoLab2$trial(example_pop, rep = 3, offset = c(-5, 0))
pheno2$cost
summary(pheno2$data)

pheno3 <- phenoLab2$trial(example_pop,
                          rep = round(runif(example_pop$nInd, 1, 3)),
                          offset = c(-5, 0))
pheno3$cost
summary(pheno3$data)

## ------------------------------------------------
## Method `phenotyper$he`
## ------------------------------------------------

phenoLab1$he(example_pop)
phenoLab2$he(example_pop)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{[character] name of the phenotyper}

\item{\code{plotCost}}{[numeric] cost for phenotyping one plot}

\item{\code{traits}}{[list] of phenotyped traits}

\item{\code{mu}}{[numeric] vector of the mean for each trait}

\item{\code{ve}}{[numeric] vector of the environmental variance sigma^2
for each trait}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{phenotyper$new()}}
\item \href{#method-print}{\code{phenotyper$print()}}
\item \href{#method-trial}{\code{phenotyper$trial()}}
\item \href{#method-he}{\code{phenotyper$he()}}
\item \href{#method-clone}{\code{phenotyper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new phenotyper object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{phenotyper$new(
  name,
  traits,
  plotCost = 1,
  mu = 0,
  ve = NULL,
  he = NULL,
  pop = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{[character] name of the phenotyper}

\item{\code{traits}}{[trait or list] of phenotyped traits}

\item{\code{plotCost}}{[numeric] cost for phenotyping one plot}

\item{\code{mu}}{[numeric] vector of the mean for each trait}

\item{\code{ve}}{[numeric] (if \code{he} not provided) vector of the
environmental variance sigma^2 for each trait}

\item{\code{he}}{[numeric] (if \code{ve} not provided) vector of the target
heredity for each trait}

\item{\code{pop}}{[population] (if \code{ve} not provided) population used for
the target heredity calculation (see:\link[breedSimulatR]{population})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `phenotyper` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySpec <- specie$new(nChr = 10,
                    lchr = 10^6,
                    lchrCm = 100,
                    specName = "Geneticae Exempli",
                    ploidy = 2)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                   specie = mySpec)
example_pop <- createPop(geno = exampleData$genotypes,
                        SNPinfo = SNPs,
                        popName = "Example population")
myTrait1 <- trait$new(name = "Trait1",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.3))
myTrait2 <- trait$new(name = "Trait2",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.4))

phenoLab1 <- phenotyper$new(name = "My phenoLab",
                            traits = list(myTrait1, myTrait2),
                            plotCost = 1,
                            mu = c(100, 50),
                            ve = NULL,
                            he = c(0.4, 0.55),
                            pop = example_pop)

phenoLab2 <- phenotyper$new(name = "My phenoLab",
                           traits = list(myTrait1, myTrait2),
                           plotCost = 2,
                           mu = c(100, 50),
                           ve = c(9, 7),
                           he = NULL,
                           pop = NULL)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display informations about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{phenotyper$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trial"></a>}}
\if{latex}{\out{\hypertarget{method-trial}{}}}
\subsection{Method \code{trial()}}{
Phenotype a given population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{phenotyper$trial(pop, rep = 1, offset = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{[population class] population
(see: \link[breedSimulatR]{population})}

\item{\code{rep}}{[numeric] number of replication for each individuals}

\item{\code{offset}}{[numeric] offset added to the phenotypic calculation
results}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
phenotypic values for individual i repetition j is calculated as follow:
\deqn{y_{ij} = \mu + g_i + e_ij + offset}
\deqn{e_ij \sim N(0, \sigma_e^2)}
where g_i is the genetic value of the individual i
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
pheno1 <- phenoLab1$trial(example_pop, rep = 4, offset = c(3, 5))
pheno1$cost
summary(pheno1$data)

pheno2 <- phenoLab2$trial(example_pop, rep = 3, offset = c(-5, 0))
pheno2$cost
summary(pheno2$data)

pheno3 <- phenoLab2$trial(example_pop,
                          rep = round(runif(example_pop$nInd, 1, 3)),
                          offset = c(-5, 0))
pheno3$cost
summary(pheno3$data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-he"></a>}}
\if{latex}{\out{\hypertarget{method-he}{}}}
\subsection{Method \code{he()}}{
Calculate the heritability of a given population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{phenotyper$he(pop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{[population] the population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
named numeric vector of the calculated heritability of the
population for each trait
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{phenoLab1$he(example_pop)
phenoLab2$he(example_pop)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{phenotyper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
