% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{An R6 class representing a population}
\description{
population object store specific information about one individual
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{[string] Name of the population}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[breedSimulatR]{specie})}

\item{\code{inds}}{[list] list of population's individuals}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{inds}}{[list] list of population's individuals}

\item{\code{nInd}}{[numeric] number of individual in the population}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{population$new()}}
\item \href{#method-addInds}{\code{population$addInds()}}
\item \href{#method-remInds}{\code{population$remInds()}}
\item \href{#method-print}{\code{population$print()}}
\item \href{#method-clone}{\code{population$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new population object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$new(name = "New Pop", inds = list(), verbose = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{[string] name of the population}

\item{\code{inds}}{[individual class or list] list of individuals of the
population (see:\link[breedSimulatR]{individual})}

\item{\code{verbose}}{[boolean] display information}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `population` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addInds"></a>}}
\subsection{Method \code{addInds()}}{
Add individuals to the population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$addInds(inds, verbose = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inds}}{[individual class or list] list of individuals of the
population (see:\link[breedSimulatR]{individual})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remInds"></a>}}
\subsection{Method \code{remInds()}}{
Remove individuals from the population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$remInds(indsNames, verbose = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indsNames}}{[character] character vercor of the individuals' names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{population$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
