% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mating.R
\name{randomMate}
\alias{randomMate}
\title{Random Mating}
\usage{
randomMate(inds, n, names)
}
\arguments{
\item{inds}{Vector of the selected individuals' names}

\item{n}{Total number of cross to perform}

\item{names}{Names of the new individuals.}
}
\value{
\code{data.frame} of the crossing table
}
\description{
Random Mating
}
\examples{
nCross <- 10
inds <- paste0("ind-", seq(1:5))
names <- paste0("newInd-", seq(1:nCross))
randomMate(inds, nCross, names)
}
