% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readVCF}
\alias{readVCF}
\title{Create `population` and `SNPinfo` objects from VCF file}
\usage{
readVCF(file, specie = NULL, verbose = TRUE)
}
\arguments{
\item{file}{path of a `.vcf` or `.vcf.gz` file.}

\item{specie}{[specie class] Specie of the population
(see:\link[breedSimulatR]{specie}), if this parameter is not provided the function will create one based on the information in the VCF file. See details section for more information.}

\item{verbose}{If TRUE, display information on the function progress}
}
\value{
list of 3 elements `pop` of class `population`, `snps` of class `snpInfo` and `specie` of class specie
}
\description{
Create `population` and `SNPinfo` objects from VCF file
}
\details{
If `specie` is not provided, the function will create one based on the information in the VCF file. The number of chromosome will be set to the number of differents chromosome in the VCF file, the names of the chromosomes will be set the the unique values of the column `CHROM` of the VCF file, the physical length of each chromosome will equal to the highest physical position of the chromosome's markers, the length of chromosomes in centimorgans will be set to 1cm per Mega Base.
}
\examples{
\dontrun{
# create vcf.gz file
mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     lchrCm = 100,
                     specName = "Geneticae Exempli")
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
vcfFile <- tempfile(fileext = "vcf.gz") 
example_pop$writeVcf(vcfFile)

# read vcf file
x <- readVCF(file = vcfFile , verbose = TRUE)
print(x$pop)
print(x$snps)
print(x$specie)
}
}
