% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selectWBV}
\alias{selectWBV}
\title{Selection according to weighted breeding values}
\usage{
selectWBV(pop, n, SNPeffects)
}
\arguments{
\item{pop}{(population object) Population of individuals to select
(see:\link[breedSimulatR]{population})}

\item{n}{(integer) number of individuals to select}

\item{SNPeffects}{(numeric vector) effect of the genetic markers}
}
\value{
character vector of the selected individuals' names
}
\description{
Selection according to weighted breeding values
}
\examples{
mySpec <- specie$new(specName = "Statisticae exempli",
                     nChr = 10,
                     lchr = 1e6,
                     ploidy = 2,
                     recombRate = 3/1e6)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord, specie = mySpec)
initPop <- createPop(geno = exampleData$genotypes,
                     SNPinfo = SNPs,
                     popName = "Initial population")
selectWBV(pop = initPop,
         SNPeffects = exampleData$snpEffects,
         n = 10)
}
\references{
Jannink, Jean-Luc. “Dynamics of Long-Term Genomic Selection.”
  Genetics Selection Evolution 42, no. 1 (December 2010).
  https://doi.org/10.1186/1297-9686-42-35.
}
