% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP.R
\name{SNPinfo}
\alias{SNPinfo}
\title{R6 Class representing a set of SNP markers}
\description{
SNPinfo object store specific information of a set of SNP
markers
}
\details{
This class is useful for setting the Haplotype class. (see:
\code{\link[breedSimulatR]{haplotype}})
}
\examples{

## ------------------------------------------------
## Method `SNPinfo$new`
## ------------------------------------------------

# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)

## ------------------------------------------------
## Method `SNPinfo$nSNP`
## ------------------------------------------------

SNPs$nSNP()
SNPs$nSNP(c("Chr2","Chr3"))

## ------------------------------------------------
## Method `SNPinfo$getInfo`
## ------------------------------------------------

SNPs$getInfo("SNP01")
SNPs$getInfo(c("SNP01", "SNP03"))

## ------------------------------------------------
## Method `SNPinfo$plot`
## ------------------------------------------------

SNPs$plot(alpha = 1)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SNPcoord}}{[data.frame] Coordinate of all SNPs.

3 columns:
\itemize{
 \item{\code{chr}:} {Chromosome holding the SNP}
 \item{\code{pos}:} {SNP position on the chromosome}
 \item{\code{SNPid}:} {SNP's IDs}
 }}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[breedSimulatR]{specie})}

\item{\code{SNPcoordList}}{[list] Named list of dataframes with the coordinate
of all SNPs. for each chromosomes}

\item{\code{ids}}{[list] Named list of the SNP ids for all chromosomes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SNPinfo$new()}}
\item \href{#method-nSNP}{\code{SNPinfo$nSNP()}}
\item \href{#method-getInfo}{\code{SNPinfo$getInfo()}}
\item \href{#method-print}{\code{SNPinfo$print()}}
\item \href{#method-plot}{\code{SNPinfo$plot()}}
\item \href{#method-clone}{\code{SNPinfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SNPinfo object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$new(SNPcoord, specie)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SNPcoord}}{[data.frame] Coordinate of all SNPs.

  3 columns: \itemize{ \item{\code{chr}:} {Chromosome holding the SNP}
  \item{\code{pos}:} {SNP position on the chromosome}
  \item{\code{SNPid}:} {SNP's IDs} }}

\item{\code{specie}}{[specie class] Specie of the SNPs
(see:\link[breedSimulatR]{specie})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SNPinfo` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# create specie
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     verbose = FALSE)

# simulate SNP
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nSNP"></a>}}
\if{latex}{\out{\hypertarget{method-nSNP}{}}}
\subsection{Method \code{nSNP()}}{
Get the number of SNPs per chromosomes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$nSNP(chr = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chr}}{[str or numeric] chromosome id}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{SNPs$nSNP()
SNPs$nSNP(c("Chr2","Chr3"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getInfo}{}}}
\subsection{Method \code{getInfo()}}{
Get information about specific SNPs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$getInfo(SNPid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SNPid}}{[str] SNP ids}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{SNPs$getInfo("SNP01")
SNPs$getInfo(c("SNP01", "SNP03"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display summary information about the object: specie, number
of SNP, SNP coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
plot chromosome map using the \pkg{plotly} package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$plot(alpha = 0.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{transparency see \link[plotly]{plot_ly}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{SNPs$plot(alpha = 1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SNPinfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
