% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specie.R
\docType{data}
\name{specie}
\alias{specie}
\title{Specie class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
specie
}
\value{
Specie R6 object
}
\description{
An R6 class defining a specie.
}
\section{Class Constructor}{

\describe{
    \item{\code{new(nChr, lchr, specName = "Undefinded", ploidy = NA, mutRate = NA, recombRate = NA,}}{
        \itemize{
            \item{Generate a new specie}
            \item{\bold{Args:}}{
                \itemize{
                    \item{\bold{\code{nChr}}: [numeric] Number of chromosomes / ploidy
                 }
                    \item{\bold{\code{lchr}}: [numeric] Length of each chromosomes
                 }
                    \item{\bold{\code{specName}}: [str] specie's name
                 }
                    \item{\bold{\code{ploidy}}: [numeric] the number of possible alleles at one locus
                 }
                    \item{\bold{\code{mutRate}}: [numeric] Mutation rate at each base
                 }
                    \item{\bold{\code{recombRate}}: [numeric] Mutation rate at each base
                 }
                    \item{\bold{\code{chrNames}}: [str] Names of the chromosomes
                 }
                    \item{\bold{\code{verbose}}: [bool]Display info
                 }
                }
            }
            \item{\bold{Returns:}}{
                \itemize{
                    \item{New "specie" object}
                }
            }
        }
    }
}
}

\section{Public Methods}{

\describe{
    \item{\code{getChrLength(chr)}}{
        \itemize{
            \item{Get the chromosome length}
            \item{\bold{Args:}}{
                \itemize{
                    \item{\bold{\code{chr}}: [str or numeric] chromosome ids
                 }
                }
            }
            \item{\bold{Returns:}}{
                \itemize{
                    \item{Named vector of the chromosomes length}
                }
            }
        }
    }
}
}

\section{Public Fields}{

\describe{
    \item{\bold{\code{specName}}}{: [str] Specie's name}
    \item{\bold{\code{nChr}}}{: [numeric] Number of chromosomes}
    \item{\bold{\code{ploidy}}}{: [numeric] Number of possible alleles at one locus}
    \item{\bold{\code{lchr}}}{: [numeric] length of all chromosomes}
    \item{\bold{\code{mutRate}}}{: [numeric] Mutation rate at each base}
    \item{\bold{\code{chrNames}}}{: [str] Names of the chromosomes}
    \item{\bold{\code{recombRate}}}{: [numeric] Mutation rate at each base}
}
}

\section{Special Methods}{

\describe{
    \item{\code{clone(deep = FALSE)}}{
        \itemize{
            \item{Method for copying an object. See \href{https://adv-r.hadley.nz/r6.html#r6-semantics}{emph{Advanced R}} for the intricacies of R6 reference semantics.}
            \item{\bold{Args:}}{
                \itemize{
                    \item{\bold{\code{deep}}: logical. Whether to recursively clone nested R6 objects.
                 }
                }
            }
            \item{\bold{Returns:}}{
                \itemize{
                    \item{Cloned object of this class.}
                }
            }
        }
    }
    \item{\code{print()}}{
        \itemize{
            \item{Display information about the object}
        }
    }
}
}

\examples{
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     specName = "Geneticae Exempulus",
                     ploidy = 2,
                     mutRate = 10^-8,
                     recombRate = 10^-7)
print(mySpec)

}
\keyword{datasets}
