% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{trait}
\alias{trait}
\title{R6 class representing a phenotypic trait}
\description{
trait object store information about a phenotypic trait
}
\examples{

## ------------------------------------------------
## Method `trait$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     specName = "Geneticae Exempli",
                     ploidy = 2,
                     recombRate = 3/10^6)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     class = "quantitative",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))

## ------------------------------------------------
## Method `trait$gv`
## ------------------------------------------------

# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{[string] Name of the trait}

\item{\code{class}}{"qualitative" or "quantitative"}

\item{\code{arch}}{Genetic architecture}

\item{\code{qtn}}{[character vector] list of the quantitative trait nucleotides}

\item{\code{qtnEff}}{[character vector] quantitative trait nucleotides effects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trait$new()}}
\item \href{#method-gv}{\code{trait$gv()}}
\item \href{#method-print}{\code{trait$print()}}
\item \href{#method-clone}{\code{trait$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new trait object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trait$new(
  name = NULL,
  class = "quantitative",
  arch = "additive",
  qtn = NULL,
  qtnEff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{[] name description}

\item{\code{class}}{"quantitative" or "qualitative"}

\item{\code{arch}}{"arch Genetic architecture}

\item{\code{qtn}}{qtn [character vector] list of the quantitative trait
nucleotides}

\item{\code{qtnEff}}{qtnEff [numeric vector] quantitative trait nucleotides
effects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\code{qtn} and \code{qtnEff} must be in the same order.
  \code{qtnEff[n]} must be the QTN effect of the QTN \code{qtn[n]}.
}

\subsection{Returns}{
A new `trait` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySpec <- specie$new(nChr = 10,
                     lchr = 10^6,
                     specName = "Geneticae Exempli",
                     ploidy = 2,
                     recombRate = 3/10^6)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
                    specie = mySpec)

myTrait <- trait$new(name = "myTrait",
                     class = "quantitative",
                     qtn = sample(SNPs$SNPcoord$SNPid, 100),
                     qtnEff = rnorm(100, sd = 0.5))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gv"></a>}}
\if{latex}{\out{\hypertarget{method-gv}{}}}
\subsection{Method \code{gv()}}{
Calculate the genetic values of a population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trait$gv(pop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pop}}{[population class] population
(see:\link[breedSimulatR]{population})}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# create population
example_pop <- createPop(geno = exampleData$genotypes,
                         SNPinfo = SNPs,
                         popName = "Example population")
myTrait$gv(example_pop)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trait$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trait$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
