% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedSimulatR-package.R
\docType{package}
\name{breedSimulatR-package}
\alias{breedSimulatR}
\alias{breedSimulatR-package}
\title{breedSimulatR: R Breeding Simulator}
\description{
R package providing classes and functions to
    simulate breeding schemes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ut-biomet/breedSimulatR}
  \item Report bugs at \url{https://github.com/ut-biomet/breedSimulatR/issues}
}

}
\author{
\strong{Maintainer}: Julien Diot \email{juliendiot@ut-biomet.org}

Authors:
\itemize{
  \item Hiroyoshi Iwata \email{aiwata@mail.ecc.u-tokyo.ac.jp} [contributor]
}

}
\keyword{internal}
