% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{makeCrosses}
\alias{makeCrosses}
\title{Proceed to several crosses}
\usage{
makeCrosses(crosses, pop)
}
\arguments{
\item{crosses}{data.frame with crossing instructions: parents names
\code{ind1} \code{ind2}, number of descendents \code{n} and names of
descendents \code{names}}

\item{pop}{list of individuals containing the parents}
}
\value{
list of new individuals
}
\description{
Proceed to several crosses
}
