% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{makeSingleCross}
\alias{makeSingleCross}
\title{Cross two individuals together}
\usage{
makeSingleCross(ind1, ind2, n = 1, names = NULL, verbose = T)
}
\arguments{
\item{ind1}{parent 1}

\item{ind2}{parent 2}

\item{n}{number of descendants}

\item{names}{names of the descendants}

\item{verbose}{print informations}
}
\value{
list of new individuals
}
\description{
Cross two individuals together
}
\examples{
# create specie:
mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     recombRate = 0.006,
                     verbose = FALSE)

# simulate SNP:
SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
                               rep("Chr2", 4),
                               rep("Chr3", 5)),
                       pos = c(sample(100, 3),
                               sample(150, 4),
                               sample(200, 5)),
                       SNPid = sprintf(fmt = paste0("SNP\%0", 2,"i"),
                                       1:(3 + 4 + 5)))

# create SNPinfo object:
SNPs <- SNPinfo$new(SNPcoord = SNPcoord, specie = mySpec)


# simulate haplotype:
rawHaplo1 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE), nrow = 2)
colnames(rawHaplo1) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo1 <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo1)
rawHaplo2 <- matrix(sample(c(0, 1), (3 + 4 + 5) * 2, replace = TRUE), nrow = 2)
colnames(rawHaplo2) <- sprintf(fmt = paste0("SNP\%0", 2,"i"),
                              1:(3 + 4 + 5))
haplo2 <- haplotype$new(SNPinfo = SNPs,
                       haplo = rawHaplo2)


# create individuals:
myInd1 <-  individual$new(name = "Ind 1",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo1,
                         verbose = FALSE)
myInd2 <-  individual$new(name = "Ind 2",
                         specie = mySpec,
                         parent1 = "OkaaSan",
                         parent2 = "OtouSan",
                         haplo = haplo2,
                         verbose = FALSE)
offspring <- makeSingleCross(myInd1, myInd2, names = "off 1")
offspring
}
