% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selectBV}
\alias{selectBV}
\title{Selection according to breeding values}
\usage{
selectBV(pop, SNPeffects, n)
}
\arguments{
\item{pop}{(population object) Population of individuals to select
(see:\link[breedSimulatR]{population})}

\item{SNPeffects}{(numeric vector) effect of the genetic markers}

\item{n}{(integer) number of individuals to select}
}
\value{
character vector of the selected individulas' names
}
\description{
Selection according to breeding values
}
\examples{
mySpec <- specie$new(specName = "Statisticae exempli",
                     nChr = 10,
                     lchr = 1e6,
                     ploidy = 2,
                     recombRate = 3/1e6)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord, specie = mySpec)
initPop <- createPop(geno = exampleData$genotypes,
                     SNPinfo = SNPs,
                     popName = "Initial population")
selectBV(pop = initPop,
         SNPeffects = exampleData$snpEffects,
         n = 10)
}
