% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selectBV}
\alias{selectBV}
\title{Selection according to breeding values}
\usage{
selectBV(pop, n, QTNeffects)
}
\arguments{
\item{pop}{(population object) Population of individuals to select
(see:\link[breedSimulatR]{population})}

\item{n}{(integer) number of individuals to select}

\item{QTNeffects}{(named numeric vector) effect of the genetic markers}
}
\value{
character vector of the selected individuals' names
}
\description{
Selection according to breeding values
}
\examples{
mySpec <- specie$new(specName = "Statisticae exempli",
                     nChr = 10,
                     lchr = 1e6,
                     lchrCm = 100)
SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord, specie = mySpec)
initPop <- createPop(geno = exampleData$genotypes,
                     SNPinfo = SNPs,
                     popName = "Initial population")
myTrait <- trait$new(name = "myTrait",
                     qtn = names(exampleData$snpEffects),
                     qtnEff = exampleData$snpEffects)
selectBV(pop = initPop,
         QTNeffects = myTrait$qtnEff,
         n = 10)
}
