% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specie.R
\name{specie}
\alias{specie}
\title{R6 Class Representing a Specie}
\description{
Specie object store specific information of one specie.
}
\examples{

## ------------------------------------------------
## Method `specie$new`
## ------------------------------------------------

mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     lchrCm = 100,
                     specName = "Geneticae Exempulus")
print(mySpec)

## ------------------------------------------------
## Method `specie$getChrLength`
## ------------------------------------------------

mySpec$getChrLength()
mySpec$getChrLength(2)
mySpec$getChrLength("Chr3")

## ------------------------------------------------
## Method `specie$getChrLengthCm`
## ------------------------------------------------

mySpec$getChrLengthCm()
mySpec$getChrLengthCm(2)
mySpec$getChrLengthCm("Chr3")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{specName}}{[str] Specie's name}

\item{\code{nChr}}{[numeric] Number of chromosomes}

\item{\code{ploidy}}{[numeric] Number of possible alleles at one locus}

\item{\code{lchr}}{[numeric] length of all chromosomes in base pairs}

\item{\code{lchrCm}}{[numeric] length of all chromosomes in centimorgans}

\item{\code{mutRate}}{[numeric] Mutation rate at each base}

\item{\code{chrNames}}{[str] Names of the chromosomes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{specie$new()}}
\item \href{#method-print}{\code{specie$print()}}
\item \href{#method-getChrLength}{\code{specie$getChrLength()}}
\item \href{#method-getChrLengthCm}{\code{specie$getChrLengthCm()}}
\item \href{#method-clone}{\code{specie$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new specie object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specie$new(
  nChr,
  lchr,
  lchrCm,
  specName = "Undefinded",
  ploidy = NA,
  mutRate = NA,
  chrNames = NA,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nChr}}{[numeric] Number of chromosomes}

\item{\code{lchr}}{[numeric] length of all chromosomes in base pairs}

\item{\code{lchrCm}}{[numeric] length of all chromosomes in centimorgans}

\item{\code{specName}}{[str] Specie's name (optional)}

\item{\code{ploidy}}{[numeric] Number of possible alleles at one locus
(optional)}

\item{\code{mutRate}}{[numeric] Mutation rate at each base (optional)}

\item{\code{chrNames}}{[str] Names of the chromosomes (optional)}

\item{\code{verbose}}{[bool] Display info (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `specie` object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySpec <- specie$new(nChr = 3,
                     lchr = c(100, 150, 200),
                     lchrCm = 100,
                     specName = "Geneticae Exempulus")
print(mySpec)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Display information about the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specie$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChrLength"></a>}}
\if{latex}{\out{\hypertarget{method-getChrLength}{}}}
\subsection{Method \code{getChrLength()}}{
Get the chromosomes length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specie$getChrLength(chr = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chr}}{[str or numeric] chromosome ids}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySpec$getChrLength()
mySpec$getChrLength(2)
mySpec$getChrLength("Chr3")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChrLengthCm"></a>}}
\if{latex}{\out{\hypertarget{method-getChrLengthCm}{}}}
\subsection{Method \code{getChrLengthCm()}}{
Get the chromosomes length in centimorgans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specie$getChrLengthCm(chr = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chr}}{[str or numeric] chromosome ids}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mySpec$getChrLengthCm()
mySpec$getChrLengthCm(2)
mySpec$getChrLengthCm("Chr3")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{specie$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
